refactor define pg polyconstraint
